+++ src/api2/index.js
@@ -1,8 +1,10 @@
 export { ApiCaller } from './api-caller'
 export {
+  VerificationManager,
   DocumentsManager,
   FactorsManager,
   SignersManager,
-  WalletsManager
+  WalletsManager,
+  Signer
 } from './managers'
 export { JsonapiResponse } from './middlewares/parse-jsonapi-response'